+++ adventure-test/src/paginator.rs
@@ -1,11 +1,14 @@
 use std::marker::PhantomData;
+use std::pin::Pin;
 use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::Arc;
 
 use adventure::{
     oneshot::OneshotRequest,
     paginator::PagedRequest,
     request::{BaseRequest, Request},
 };
+use pin_utils::pin_mut;
 
 struct MockClient<T> {
     called: AtomicUsize,
@@ -25,34 +28,45 @@ impl<T> MockClient<T> {
 
 #[derive(Debug, Default)]
 struct Numbers {
-    current: AtomicUsize,
+    current: Arc<AtomicUsize>,
     end: usize,
+}
+
+impl Numbers {
+    fn new(start: usize, end: usize) -> Self {
+        Numbers {
+            current: Arc::new(AtomicUsize::from(start)),
+            end,
+        }
+    }
 }
 
 macro_rules! test_cases {
     () => {
-        impl BaseRequest for &Numbers {
+        impl BaseRequest for Numbers {
             type Ok = usize;
             type Error = ();
         }
 
-        impl OneshotRequest<&MockClient<Response>> for &Numbers {
+        impl OneshotRequest<&MockClient<Response>> for Numbers {
             type Response = Response;
 
             fn send_once(self, client: &MockClient<Response>) -> Self::Response {
-                self.send(client)
+                let this = self;
+                pin_mut!(this);
+                this.send(client)
             }
         }
 
-        impl Request<&MockClient<Response>> for &Numbers {
+        impl Request<&MockClient<Response>> for Numbers {
             type Response = Response;
 
-            fn send(&self, client: &MockClient<Response>) -> Self::Response {
+            fn send(self: Pin<&mut Self>, client: &MockClient<Response>) -> Self::Response {
                 MockClient::<Response>::send_request(client, self)
             }
         }
 
-        impl PagedRequest for &Numbers {
+        impl PagedRequest for Numbers {
             fn advance(&mut self, response: &Self::Ok) -> bool {
                 if *response < self.end {
                     self.current.fetch_add(1, Ordering::SeqCst);
@@ -66,87 +80,78 @@ macro_rules! test_cases {
         #[test]
         fn paginator_basic() {
             let client = MockClient::<Response>::new(|_| true);
-            let numbers = Numbers {
-                current: AtomicUsize::new(1),
-                end: 5,
-            };
-            let paginator = (&numbers).paginate(&client);
+            let numbers = Numbers::new(1, 5);
+            let current = Arc::clone(&numbers.current);
+            let paginator = numbers.paginate(&client);
 
             let responses = collect(paginator);
             assert_eq!(Ok(vec![1, 2, 3, 4, 5]), responses);
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 5);
-            assert_eq!(numbers.end, 5);
+            assert_eq!(current.load(Ordering::SeqCst), 5);
             assert_eq!(client.called.load(Ordering::SeqCst), 5);
         }
 
         #[test]
         fn paginator_basic_2() {
             let client = MockClient::<Response>::new(|n| n.current.load(Ordering::SeqCst) < 7);
-            let numbers = Numbers {
-                current: AtomicUsize::new(1),
-                end: 20,
-            };
-            let paginator = (&numbers).paginate(&client);
+            let numbers = Numbers::new(1, 20);
+            let current = Arc::clone(&numbers.current);
+            let paginator = numbers.paginate(&client);
 
             let responses = collect(paginator);
             assert_eq!(Err(()), responses);
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 7);
+            assert_eq!(current.load(Ordering::SeqCst), 7);
             assert_eq!(client.called.load(Ordering::SeqCst), 7);
         }
 
         #[test]
         fn paginator_step() {
             let client = MockClient::<Response>::new(|_| true);
-            let numbers = Numbers {
-                current: AtomicUsize::new(1),
-                end: 3,
-            };
-            let mut paginator = Some((&numbers).paginate(&client));
+            let numbers = Numbers::new(1, 3);
+            let current = Arc::clone(&numbers.current);
+            let mut paginator = Some(numbers.paginate(&client));
 
             assert_eq!(block_on_next(&mut paginator), Some(Ok(1)));
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 2);
+            assert_eq!(current.load(Ordering::SeqCst), 2);
             assert_eq!(client.called.load(Ordering::SeqCst), 1);
 
             assert_eq!(block_on_next(&mut paginator), Some(Ok(2)));
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 3);
+            assert_eq!(current.load(Ordering::SeqCst), 3);
             assert_eq!(client.called.load(Ordering::SeqCst), 2);
 
             assert_eq!(block_on_next(&mut paginator), Some(Ok(3)));
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 3);
+            assert_eq!(current.load(Ordering::SeqCst), 3);
             assert_eq!(client.called.load(Ordering::SeqCst), 3);
 
             assert_eq!(block_on_next(&mut paginator), None);
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 3);
+            assert_eq!(current.load(Ordering::SeqCst), 3);
             assert_eq!(client.called.load(Ordering::SeqCst), 3);
 
             assert_eq!(block_on_next(&mut paginator), None);
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 3);
+            assert_eq!(current.load(Ordering::SeqCst), 3);
             assert_eq!(client.called.load(Ordering::SeqCst), 3);
         }
 
         #[test]
         fn paginator_step_with_error() {
             let client = MockClient::<Response>::new(|n| n.current.load(Ordering::SeqCst) < 3);
-            let numbers = Numbers {
-                current: AtomicUsize::new(1),
-                end: 3,
-            };
-            let mut paginator = Some((&numbers).paginate(&client));
+            let numbers = Numbers::new(1, 3);
+            let current = Arc::clone(&numbers.current);
+            let mut paginator = Some(numbers.paginate(&client));
 
             assert_eq!(block_on_next(&mut paginator), Some(Ok(1)));
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 2);
+            assert_eq!(current.load(Ordering::SeqCst), 2);
             assert_eq!(client.called.load(Ordering::SeqCst), 1);
 
             assert_eq!(block_on_next(&mut paginator), Some(Ok(2)));
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 3);
+            assert_eq!(current.load(Ordering::SeqCst), 3);
             assert_eq!(client.called.load(Ordering::SeqCst), 2);
 
             assert_eq!(block_on_next(&mut paginator), Some(Err(())));
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 3);
+            assert_eq!(current.load(Ordering::SeqCst), 3);
             assert_eq!(client.called.load(Ordering::SeqCst), 3);
 
             assert_eq!(block_on_next(&mut paginator), Some(Err(())));
-            assert_eq!(numbers.current.load(Ordering::SeqCst), 3);
+            assert_eq!(current.load(Ordering::SeqCst), 3);
             assert_eq!(client.called.load(Ordering::SeqCst), 4);
         }
     };
@@ -164,9 +169,9 @@ mod futures01 {
     pub(super) type Response = LocalFuture01ResponseObj<'static, usize, ()>;
 
     impl MockClient<Response> {
-        pub(super) fn send_request(&self, req: &Numbers) -> Response {
+        pub(super) fn send_request(&self, req: Pin<&mut Numbers>) -> Response {
             self.called.fetch_add(1, Ordering::SeqCst);
-            if (self.pred)(req) {
+            if (self.pred)(req.as_ref().get_ref()) {
                 Response::new(future::ok(req.current.load(Ordering::SeqCst)))
             } else {
                 Response::new(future::err(()))
@@ -210,9 +215,9 @@ mod std_futures {
     pub(super) type Response = LocalFutureResponseObj<'static, usize, ()>;
 
     impl MockClient<Response> {
-        pub(super) fn send_request(&self, req: &Numbers) -> Response {
+        pub(super) fn send_request(&self, req: Pin<&mut Numbers>) -> Response {
             self.called.fetch_add(1, Ordering::SeqCst);
-            if (self.pred)(req) {
+            if (self.pred)(req.as_ref().get_ref()) {
                 Response::new(future::ok(req.current.load(Ordering::SeqCst)))
             } else {
                 Response::new(future::err(()))